requirements.txt
streamlit>=1.31.0
PyPDF2>=3.0.1
faiss-cpu>=1.7.4
numpy>=1.24.0
sentence-transformers>=2.2.2
langchain>=0.1.17
langchain-core>=0.1.35
langchain-community>=0.0.30
langchain-experimental>=0.0.50
langchain-ollama>=0.1.3
langchain-groq>=0.1.4
langchain-google-genai>=0.0.12
FlagEmbedding>=1.2.0